updated project code

+++ CRUD/src/main/java/com/example/CRUD/service/CommentService.java
@@ -22,50 +22,59 @@ public class CommentService {
     private final PostRepository postRepository;
 
 
-    @Transactional
-    public Comment addComment(Long userId, Long postId, String content) {
+    public Comment addComment(String username, Long postId, String content) {
 
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("User not found"));
-
+        User user = userRepository.findByUsername(username);
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found"));
 
         Comment comment = new Comment();
-        comment.setComment(content);
         comment.setUser(user);
         comment.setPost(post);
+        comment.setComment(content);
 
         return commentRepository.save(comment);
     }
 
-    @Transactional
-    public Comment updateComment(Long commentId, Long userId, String content) {
+
+    public Comment updateComment(Long commentId, String username, String content) {
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new RuntimeException("Comment not found"));
 
-
+        if (!comment.getUser().getUsername().equals(username)) {
+            throw new RuntimeException("You are not allowed to update this comment");
+        }
 
         comment.setComment(content);
-        return comment;
+
+        return commentRepository.save(comment);
     }
 
-    @Transactional
-    public void deleteComment(Long commentId, Long userId, boolean isAdmin) {
+
+    public void deleteComment(Long commentId, String username) {
 
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new RuntimeException("Comment not found"));
 
+        boolean isCommentOwner =
+                comment.getUser().getUsername().equals(username);
 
+        boolean isPostOwner =
+                comment.getPost().getUser().getUsername().equals(username);
+
+        if (!isCommentOwner && !isPostOwner) {
+            throw new RuntimeException("You are not allowed to delete this comment");
+        }
 
         commentRepository.delete(comment);
     }
 
-    public List<Comment> getCommentsByPost(Long postId) {
 
-        return commentRepository.findByPostIdOrderByCreatedAtDesc(postId);
+    public List<Comment> getCommentsByPost(Long postId) {
+        return commentRepository.findByPostId(postId);
     }
+
 
     public long countComments(Long postId) {
         return commentRepository.countByPostId(postId);
